Add: ファイル名変更・切り取り機能を追加

+++ manager/static/file.js
@@ -1,27 +1,41 @@
 let href = [];
-const params = (new URL(document.location)).searchParams;
+const url = new URL(document.location);
+const params = url.searchParams;
 const path = params.get('p');
 const source = params.get('s');
+const cutsource = params.get('k');
 const infos = params.get('i');
 const info_success = document.getElementById('info_success');
 const info_error = document.getElementById('info_error');
 const copy = document.getElementById('copy');
-const del = document.getElementById('del');
 const copysvg = document.getElementById('copysvg');
+const cut = document.getElementById('cut');
+const cutsvg = document.getElementById('cutsvg');
+const del = document.getElementById('del');
 const delsvg = document.getElementById('delsvg');
-const confirm = document.getElementById('confirm');
-const Delete = document.querySelector('.confirm_box');
 const paste = document.getElementById('paste');
 const pastesvg = document.getElementById('pastesvg');
 const edit = document.getElementById('edit');
 const editsvg = document.getElementById('editsvg');
+const rnm = document.getElementById('rnm');
+const rnmsvg = document.getElementById('rnmsvg');
+const confirm = document.getElementById('confirm');
+const Delete = document.getElementById('delete');
+const Rename = document.getElementById('rename');
+const rename_box = document.getElementById('rename_box');
 confirm.style.transform = 'scale(0, 0)';
 copy.style.cursor = 'not-allowed';
 del.style.cursor = 'not-allowed';
 edit.style.cursor = 'not-allowed';
+rnm.style.cursor = 'not-allowed';
 copysvg.style.fill = '#a9a9a9';
 delsvg.style.fill = '#a9a9a9';
 editsvg.style.fill = '#a9a9a9';
+rnmsvg.style.fill = '#a9a9a9';
+
+var inside = false;
+
+history.replaceState('', '', String(url).replace(/&i=.*/, ''));
 
 // ファイル数の取得
 let filenumber = $('#ul').attr('file');
@@ -31,6 +45,7 @@ function sleep(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
 }
 async function success(info){
+    (new URL(window.location.href)).searchParams.delete('i');
     info_success.style.display = 'flex';
     $('#info_success_h1').text(info[1]);
     $('.info').toggleClass('active');
@@ -46,17 +61,18 @@ async function error(info){
 }
 
 if (infos !== null && infos !== ''){
-    let info = infos.split(',')
+    let info = infos.split(',');
     if (info[0] === 's'){
-        success(info)
+        success(info);
     }else{
-        error(info)
+        error(info);
     }
 }
 
-
 // 削除の確認
-function delete_file(){
+async function delete_file(){
+    Delete.style.display = 'inline';
+    await sleep(10);
     confirm.style.transform = 'scale(1, 1)';
     Delete.classList.toggle('active');
 }
@@ -67,20 +83,55 @@ function delete_file_confirm(){
     location.href='fd?p=' + str;
 }
 
+async function rename_file(){
+    rename_box.value = $(`li[dir="${href[0]}"]`).attr('name');
+    Rename.style.display = 'inline';
+    await sleep(10);
+    confirm.style.transform = 'scale(1, 1)';
+    Rename.classList.toggle('active');
+}
+
+function rename_file_confirm(){
+    if (path.endsWith('/')){
+        to_name = `${path}${rename_box.value}`;
+    }else{
+        to_name = `${path}/${rename_box.value}`;
+    }
+    location.href = `fr?p=${href[0]}&d=${to_name}`;
+}
+
+async function inside_cancel(){
+    inside = true;
+    await sleep(101);
+    inside = false;
+}
 // 削除キャンセル
-async function cancel(){
-    Delete.classList.toggle('active');
-    await sleep(100);
-    confirm.style.transform = 'scale(0, 0)';
+async function cancel(t){
+    if (inside === false){
+        document.querySelector('.active').classList.toggle('active');
+        await sleep(100);
+        confirm.style.transform = 'scale(0, 0)';
+        Delete.style.display = 'none';
+        Rename.style.display = 'none';
+    }
 }
 // ファイルコピー
 function copy_file(){
     location.href = '?p=' + path + '&s=' + href.join(',');
 }
 
+// ファイル切り取り
+function cut_file(){
+    location.href = '?p=' + path + '&k=' + href.join(',');
+}
+
 // ファイル貼り付け
 function paste_file(){
-    location.href = './fc?s=' + source + '&d=' + path;
+    if(source != null){
+        location.href = './fc?s=' + source + '&d=' + path;
+    }else{
+        location.href = './fk?s=' + cutsource + '&d=' + path;
+    }
 }
 
 // ファイル編集
@@ -97,8 +148,19 @@ function edit_file(){
 }
 
 // コピー元のソースを取得
-if(source != null){
-    CopySource = '&s=' + source;
+if(source != null || cutsource != null){
+    if(source != null){
+        CopySource = '&s=' + source;
+    }else{
+        CopySource = '&k=' + cutsource;
+        for (let i of cutsource.split(',')){
+            try{
+                document.querySelector(`li[dir="${i}"]`).querySelector('svg').style.fill = '#969696';
+            }catch (e){
+                continue
+            }
+        }
+    }
     paste.style.cursor = 'pointer';
     pastesvg.style.fill = '#123767';
     paste.setAttribute('onclick', 'paste_file()');
@@ -135,12 +197,16 @@ $('li').click(function() {
         copy.style.cursor = 'not-allowed';
         del.style.cursor = 'not-allowed';
         edit.style.cursor = 'not-allowed';
+        rnm.style.cursor = 'not-allowed';
         copysvg.style.fill = '#a9a9a9';
         delsvg.style.fill = '#a9a9a9';
         editsvg.style.fill = '#a9a9a9';
+        rnmsvg.style.fill = '#a9a9a9';
         del.setAttribute('onclick', '');
         copy.setAttribute('onclick', '');
+        cut.setAttribute('onclick', '');
         edit.setAttribute('onclick', '');
+        rnm.setAttribute('onclick', '');
         $('#selecting').removeClass('hidden');
 
     // 選択があるとき
@@ -148,9 +214,11 @@ $('li').click(function() {
         copy.style.cursor = 'pointer';
         del.style.cursor = 'pointer';
         copysvg.style.fill = '#123767';
+        cutsvg.style.fill = '#123767';
         delsvg.style.fill = '#123767';
         del.setAttribute('onclick', 'delete_file()');
         copy.setAttribute('onclick', 'copy_file()');
+        cut.setAttribute('onclick', 'cut_file()');
 
         // 選択したものがファイルだったら
         if ($('.selected').attr('filetype') !== 'folder') {
@@ -164,6 +232,15 @@ $('li').click(function() {
         }
         $('#selecting').addClass('hidden');
         $('#selecting-number').text(href.length);
+    }
+    if(href.length === 1){
+        rnm.style.cursor = 'pointer';
+        rnm.setAttribute('onclick', 'rename_file()');
+        rnmsvg.style.fill = '#123767';
+    }else{
+        rnm.style.cursor = 'not-allowed';
+        rnmsvg.style.fill = '#a9a9a9';
+        rnm.setAttribute('onclick', '');
     }
 });
 
@@ -181,6 +258,14 @@ document.addEventListener('keydown', function(event) {
     }
 
     if (event.key === 'Enter') {
+        if(getComputedStyle(Delete).display !== 'none'){
+            delete_file_confirm();
+            return;
+        }
+        if(getComputedStyle(Rename).display !== 'none'){
+            rename_file_confirm();
+            return;
+        }
         if(href.length !== 0){
             if ($('.selected').attr('filetype') === 'folder') {
                 dir = '?p=' + $('.selected').attr('dir');
@@ -193,6 +278,9 @@ document.addEventListener('keydown', function(event) {
     }
 
     if (event.key === 'Delete') {
+        if(getComputedStyle(Rename).display !== 'none'){
+            return;
+        }
         if(href.length !== 0){
             delete_file();
         }
